Codeup

+++ algorithm/Codeup1039.java
@@ -4,8 +4,8 @@ import java.util.Scanner;
 public class Codeup1039 {
     public static void main(String[] args){
         Scanner sc = new Scanner(System.in);
-        int a=sc.nextInt();
-        int b=sc.nextInt();
+        double a=sc.nextDouble();
+        double b=sc.nextDouble();
 
         System.out.printf("%.0f",a+b);
         sc.close();
